new rules && bit switch to color

+++ Fractol.h
@@ -6,7 +6,7 @@
 /*   By: axdubois <axdubois@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/02 23:15:14 by axdubois          #+#    #+#             */
-/*   Updated: 2026/05/24 10:22:45 by axdubois         ###   ########.fr       */
+/*   Updated: 2026/05/24 11:26:38 by axdubois         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,7 +44,7 @@ typedef struct f_fract
 	double ratio;
 	double panx;
 	double pany;
-	double color;
+	int color;
 	char   type;
 	int	   multicolor;
 } t_fract;
